[feat: sarama] Removed unnecessary instrument
